pyproject.toml
[project]
name = "ibis-profiling"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.5.0",
    "fastexcel>=0.11.5",
    "ibis-framework>=12.0.0",
    "pandas>=2.3.3",
    "polars>=1.38.1",
    "pyarrow>=23.0.1",
    "pyarrow-hotfix>=0.7",
    "pydantic>=2.12.5",
]

[tool.uv]
cache-dir = "/tmp/ibis-profiling/.uv_cache"

[tool.ruff]
line-length = 100
cache-dir = "/tmp/ibis-profiling/.ruff_cache"

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term"
cache_dir = "/tmp/ibis-profiling/.pytest_cache"
pythonpath = ["src"]

[tool.coverage.run]
data_file = "/tmp/ibis-profiling/.coverage"

[dependency-groups]
dev = [
    "faker>=40.8.0",
    "openpyxl>=3.1.5",
    "playwright>=1.58.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "xlsxwriter>=3.2.9",
    "ydata-profiling>=4.18.1",
]
